Change logic to be based off of client guess
When trying to guess the rate of individual sleeps we've got a magic number tuning problem as you can kinda see in the README.

Instead we can stay closer to TCP logic by guessing the number of clients directly and  then using that to calculate the amount of time to sleep. 

Also fixed a bug where when a rate limit event hits, all the threads on a process would pile on and update the client count when only one needed to be doing it. This somewhat solves the runaway case where every thread on the system doubles the guessed number even though it was pretty close to the correct value.