Make flap resistant.
When multiple clients are getting rate limited, we should ideally be only doubling the client with the lower sleep value. For example if one client is sleeping for 12 seconds and another for 6 seconds, it is not ideal to keep the 6 second client at that time value and double the 12 seconds to 24.

With this change we don't adjust the time value on the first rate limit, but rather we wait until it is rate limited twice. 

Without this logic we can end up with a runaway sleep value i.e. 12 becomes 24 becomes 48 becomes 96, etc. Once it is at a high enough value then the rate at which it decreases is very slow, but there is still a non-trivial chance that even after sleeping for 96 seconds, it might wake up at the exact time the 6 second client is also getting rate limited.

With this new logic, two clients will take turns being the "high" number.

For example:

![](https://www.dropbox.com/s/uj94tf9dcgahcbx/Screenshot%202019-08-26%2012.20.31.png?raw=1)


You can see one client go "high" and then gradually decrease over time. The other client goes up and down a few times until it eventually goes higher than the first client. This behavior is desired so you don't have one dyno/resource "stuck" at a really high throttling rate. Ideally you want the values to spread around.

The major downside of this approach is that it means when a client is hitting it's rate limit it may hit the server many more times than is necessary. 

For 2 processes with 5 threads each, the clients were rate limited by the server a total of 82 times. Assuming this number could be extrapolated that's 41 times per process, so on a 90 dynos that would be 3,690 rate limited requests per hour, 3,690 and 88,560 per day. Which does seem high.

One option to decrease the rate of the number of rate limited requests is to decrease the rate that the sleep value is decremented. In theory if we multiplied by 0.5 then we would be at 44K rate limits (etc.)

One other issue with how we are decrementing (on every request) is that when you have a discrepancy between two processes such as one sleeping for 12 and one sleeping for 4, the process sleeping for 4 will decrease at about 3x faster because it is sleeping for 1/3 the time. We could perhaps make the decrementing logic based off of some other metric not related to the number of requests such as time.

Another option, which I'm trying is to decrease the decrement rate by squaring it