Square decrement rate
When there are 4500 requests remaining then 4500/4500 = 1 and and 1^2 is 1 so at the high end we're decrementing just as fast, but on the low end, say when there are 100 requests left, previously we would decrement by 0.02 but now we would decrement by 0.00049 which is MUCH slower.  

This does mean that clients right after their sleep values get doubled will take far longer to approach the "correct" value. It should ideally result in fewer 429 status results as it will spend a longer time decrementing.

Prior to this change we estimated that there would be 80K rate-limit events in a hypothetical scenario with 90 servers and 5 threads each. With this change there would be 23,760 events, or about 70% fewer. Which is much better, but we could probably do better.

The issue we're having is that while we want the rates to decrement slowly towards the end, the algorithm tends to have it linger in a state of getting rate limited, but not enough so that two consecutive requests trigger a multiplicative increase. 

One option could be to add code that tells clients to decrement when they've hit a rate limit event so they trigger two consecutive requests faster.

The question though, is how much to decrease it?