Revert decrement logic
While good for small sleep values, this decrement logic means that large values take FOREVER to come down since they are also firing at a reduced rate.

This value still isn't good for large values either and will allow for one runaway "fast" process to consume the bulk of resources.

We need to balance two things: When there are few clients the sleep value 
needs to be able to tend to zero in a reasonably fast manner.




If there are lots of clients then we have to account for the eventuality that one of the "slow" clients will continue to get slower while a "fast" client continues to cause rate limiting events. 

Here is an example of the runaway slow behavior I would like to try to prevent: 

![](https://www.dropbox.com/s/8pjle2rch7a71v8/Screenshot%202019-08-27%2009.03.17.png?raw=1)


Note how process four has is sleeping for nearly 250 seconds but process five is hovering around 20. It's virtually impossible for a client that fires every 250 seconds to ever fully "come down" to 20 seconds.

On one hand the system as a whole is still working, rate limit events account for about 2% of overall requests which I think is probably acceptable, but it seems that we should be able to do better.


In this case we could perhaps set a goal such as every hour clients that are sleeping more than 10 seconds should be reducing their rate so that they come down by half, so if a process is sleeping for 80 seconds, by the next hour it will be sleeping for 45 etc. 

Open questions remain though: What is a good target? Should we be shooting for 1 hour intervals? Should we be shooting for a 50% rate reduction or some static goal such as getting to 10 seconds every 6 hours. Should we change logic based on some magic threshold number (such as 10 seconds as presented here) or should we try to have an algorithm that works okay for all values (not sure what that would look like).