Switch from rate limit count to time based approach
With rate limit count, we have to account for clients that might run indefinitely. In that case the term would grow indefinitely and it would grow to dominate our rate reduction calculation.

This new approach offers 2 solutions, first is to use a term that naturally resets but acts as a proxy for flappyness. In our case a flappy client will be rate limited more often and therefor spend more time with a lower time since it's last rate limit. The second is to bound this term using an exponential decay so the Lowest that the term can go is 1.