Don't start throttling until we hit a rate limit
Since we don't know what the client code is doing, it might hit the API very infrequently and expect the code to fire instantaneously. This code assumes that until a rate limit event is triggered that the client is used rarely and does not need to sleep any before firing the request. This preserves existing behavior where as soon as a client request generated from code it fires right away. 

Once a rate limit event is triggered then the client assumes that it needs to be throttled and will begin sleep behavior.